requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asttokens==2.4.1
attrs==23.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
ctransformers==0.2.5
dataclasses-json==0.5.14
debugpy==1.8.2
decorator==5.1.1
executing==2.0.1
filelock==3.15.4
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.6.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.1
grpcio-tools==1.65.1
huggingface-hub==0.24.2
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.23
langchain-pinecone==0.1.3
langchain-text-splitters==0.2.2
langchainplus-sdk==0.0.20
langsmith==0.1.93
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numexpr==2.10.1
numpy==1.26.4
openapi-schema-pydantic==1.2.4
orjson==3.10.6
packaging==24.1
parso==0.8.4
pillow==10.4.0
pinecone-client==5.0.0
pinecone-plugin-inference==1.0.2
pinecone-plugin-interface==0.0.7
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==5.27.2
protoc-gen-openapiv2==0.0.1
psutil==6.0.0
pure_eval==0.2.3
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.7.24
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==2.2.2
sentencepiece==0.2.0
setuptools==71.1.0
six==1.16.0
SQLAlchemy==2.0.31
stack-data==0.6.3
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.43.2
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.3
yarl==1.9.4
-e .